Renamed classes.
