+++ demos/demo_metal_lens.py
@@ -27,7 +27,7 @@ camera.ray_extinction_prob = 0.01
 camera.spectral_rays = 1
 camera.spectral_samples = 20
 camera.pixels = (360, 240) #(720, 480)
-camera.pixel_samples = 50
+camera.pixel_samples = 200
 camera.display_progress = True
 camera.display_update_time = 10
 camera.accumulate = True
